requirements.txt
openai
python-dotenv
langchain
pydantic